refactor(prom): Outline `ResponseFuture` into a submodule
this is a bit of a self-contained system, with `ResponseFuture` acting
as the public face of it. let's codify that by making it its own
submodule.
